* refactor scm tools * bug fix commit message * bug fix error message new lines * add svn project
Former-commit-id: 201b56b9679f12cda2c6ea989d470503f1a86678

+++ TermGenie/TermGenieCore/src/test/java/org/bbop/termgenie/tools/TempTestFolderTools.java
@@ -74,7 +74,9 @@ public class TempTestFolderTools {
 	 */
 	public static void deleteTestFolder(File testFolder) {
 		try {
-			FileUtils.deleteDirectory(testFolder);
+			if (testFolder != null) {
+				FileUtils.deleteDirectory(testFolder);
+			}
 		} catch (IOException exception) {
 			throw new RuntimeException(exception);
 		}
